requirements.txt
streamlit
snowflake-connector-python
ollama
python-dotenv
openai
